+++ src/scrabble/bag.h
@@ -26,7 +26,7 @@ class Bag {
   const std::vector<Letter>& Letters() const { return letters_; }
   void SetLetters(const std::vector<Letter>& letters) { letters_ = letters; }
   void InsertTiles(const LetterString& tiles,
-                   const std::vector<uint64_t>& exchange_insertion_dividends,
+                   const std::vector<uint16_t>& exchange_insertion_dividends,
                    std::size_t* exchange_insertion_index);
 
  private:
